feat(ast)!: add `Program::comments` (#6445)
