+++ crates/oxc_ast/src/ast/js.rs
@@ -41,6 +41,9 @@ pub struct Program<'a> {
     pub source_type: SourceType,
     #[serde(skip)]
     pub source_text: &'a str,
+    /// Sorted comments
+    #[serde(skip)]
+    pub comments: Vec<'a, Comment>,
     pub hashbang: Option<Hashbang<'a>>,
     pub directives: Vec<'a, Directive<'a>>,
     pub body: Vec<'a, Statement<'a>>,
