+++ crates/oxc_ast/src/lib.rs
@@ -57,10 +57,11 @@ pub use generated::{ast_builder, ast_kind};
 pub use num_bigint::BigUint;
 
 pub use crate::{
+    ast::comment::{Comment, CommentKind, CommentPosition},
     ast_builder::AstBuilder,
     ast_builder_impl::NONE,
     ast_kind::{AstKind, AstType},
-    trivia::{Comment, CommentKind, CommentPosition, SortedComments, Trivias},
+    trivia::{SortedComments, Trivias},
     visit::{Visit, VisitMut},
 };
