+++ tasks/ast_tools/src/main.rs
@@ -31,6 +31,7 @@ static SOURCE_PATHS: &[&str] = &[
     "crates/oxc_ast/src/ast/js.rs",
     "crates/oxc_ast/src/ast/ts.rs",
     "crates/oxc_ast/src/ast/jsx.rs",
+    "crates/oxc_ast/src/ast/comment.rs",
     "crates/oxc_syntax/src/number.rs",
     "crates/oxc_syntax/src/operator.rs",
     "crates/oxc_span/src/span/types.rs",
